requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
Django==3.2
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.13.3
django-rosetta==0.9.7
django-summernote==0.8.11.6
idna==3.2
Pillow==8.3.2
polib==1.1.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
